refactor: organize Kuru types and slippage bounds
