+++ packages/protocols/kuru/src/index.ts
@@ -1,2 +1,3 @@
 export { KuruOrderbookAbi, KuruRouterAbi } from "./abis/kuru.js";
-export { KURU_ROUTER_ADDRESS, Kuru, type KuruSwapOutcome } from "./kuru.js";
+export { KURU_ROUTER_ADDRESS, Kuru } from "./kuru.js";
+export type { KuruSwapOutcome } from "./types.js";
